+++ src/sll/api/_generated_raw.c
@@ -2886,6 +2886,45 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_path_list_dir_raw(sll_ob
 
 
 
+__SLL_API_TYPE_sll_api_path_mkdir sll_api_path_mkdir(__SLL_API_ARGS_sll_api_path_mkdir);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_path_mkdir_raw(sll_object_t*const* al,sll_arg_count_t all){
+	sll_object_t* a=NULL;
+	if (all>0){
+		a=*(al+0);
+		if (SLL_OBJECT_GET_TYPE(a)==SLL_OBJECT_TYPE_STRING){
+			SLL_ACQUIRE(a);
+		}
+		else{
+			a=sll_operator_cast(a,sll_static_int[SLL_OBJECT_TYPE_STRING]);
+		}
+	}
+	else{
+		a=SLL_CREATE();
+		a->t=SLL_OBJECT_TYPE_STRING;
+		sll_string_create(0,&(a->dt.s));
+	}
+	sll_object_t* b=NULL;
+	if (all>1){
+		b=*(al+1);
+		if (SLL_OBJECT_GET_TYPE(b)==SLL_OBJECT_TYPE_INT){
+			SLL_ACQUIRE(b);
+		}
+		else{
+			b=sll_operator_cast(b,sll_static_int[SLL_OBJECT_TYPE_INT]);
+		}
+	}
+	else{
+		b=SLL_ACQUIRE_STATIC_INT(0);
+	}
+	sll_bool_t out=sll_api_path_mkdir(&(a->dt.s),!!(b->dt.i));
+	SLL_RELEASE(b);
+	SLL_RELEASE(a);
+	SLL_ACQUIRE(sll_static_int[out]);
+	return sll_static_int[out];
+}
+
+
+
 __SLL_API_TYPE_sll_api_path_recursive_list_dir sll_api_path_recursive_list_dir(__SLL_API_ARGS_sll_api_path_recursive_list_dir);
 __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_path_recursive_list_dir_raw(sll_object_t*const* al,sll_arg_count_t all){
 	sll_object_t* a=NULL;
@@ -5382,6 +5421,10 @@ static const internal_function_t _ifunc_data_ptr[]={
 		sll_api_path_list_dir_raw
 	},
 	{
+		SLL_CHAR("sll:path_mkdir"),
+		sll_api_path_mkdir_raw
+	},
+	{
 		SLL_CHAR("sll:path_recursive_list_dir"),
 		sll_api_path_recursive_list_dir_raw
 	},
@@ -5665,5 +5708,5 @@ static const internal_function_t _ifunc_data_ptr[]={
 
 
 
-const sll_function_index_t _ifunc_size=159;
+const sll_function_index_t _ifunc_size=160;
 const internal_function_t* _ifunc_data=(const internal_function_t*)(&_ifunc_data_ptr);