+++ src/sll/api/path_api.c
@@ -179,6 +179,15 @@ __API_FUNC(path_list_dir){
 
 
 
+__API_FUNC(path_mkdir){
+	if (sll_get_sandbox_flag(SLL_SANDBOX_FLAG_DISABLE_PATH_API)){
+		return 0;
+	}
+	return sll_platform_create_directory(a->v,b);
+}
+
+
+
 __API_FUNC(path_recursive_list_dir){
 	if (sll_get_sandbox_flag(SLL_SANDBOX_FLAG_DISABLE_PATH_API)){
 		SLL_INIT_ARRAY(out);