+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Implemented `SLL_SANDBOX_FLAG_ENABLE_FILE_DELETE`, `sll_platform_path_delete` and `file$delete`
+- Implemented `sll_platform_create_directory` and `path$mkdir`
 
 ### Changed